fix(frontend): Allow long data-layers tooltip text to wrap
I added a new class that is specific to the date layers tooltip, as the existing
class is being shared with tooltips elsewhere.

2. The max-width is an aribraty figure, chosen so that the tooltip stays at a slightly
shorter width that the collapsible component that contains the info icon.

3. I also removed a <p> tag that was wrapping the text content inside the tooltip,
as it was adding unnecessary margin that didn't look very good, and it doesn't appear
to be required anyway.

4. I also added a ternary to show a 'Loading...' message until the fetch completes,
as otherwise it would result in a very small empty tooltip before the data returns,
which also doesn't look very good. I'm not sure if this is the way to go, but it seemed
an appropriate solution to that. I would probably be fetching all of the tooltip info
data upon rendering the component, instead of a network request being made on every hover,
but don't want to be refactoring anything major before talking with the other team first.