fix(frontend): Integrate global date range filtering
This integrates a single date filter into the header, the output of which is stored in global state
and imported into every component that uses it.

When global state is initialised, and when a use manually clears the date range, the date range in
state is `null`, and therefore no date data is sent to the server, which results in unrestricted data
being returned.