fix(frontend): Create re-usable filter functions
Created a pair of re-usable filter functions.

1. `getPropertyValue`: This recursively searches for defined properties within an object,
returning the value if found, and drilling deeper if not, moving onto sibling properties after that,
and eventually returning null if no match is found.

2. `filterNodesByProperty`: This uses the above function, and takes an object of `{ property: matchString }`
params to search the target object, then filter it based on the result. The choice of a params object is to
allow unlimited properties to be simultaneously searched.

On branch SAFB-166/domain-source-filtering