fix(frontend): Refactoring of map requests polling
Just a bit of refactoring, changing some variable names etc.

Also changed the pattern a bit to match the pattern already used in the app.

This is not finished, as I'm waiting on a response to a question about polling, but this cleans
up much of the code so far.