fix(frontend): Integrate searching and timestamp display
This brings in the search bar and timestamp functionality, as it's easier to bring in here
than to try to awkwardly merge from development.

This branch is long-lived, so has some inconsistencies with development. This is due to
some tickets needing to be "paused" until other work ws completed, causing them to fall
behind a bit.

If this transferring of these two features is not done here, it would mean having to keep them
from the development branch, as well as keeping all the changes from this branch too, which
could get really messy.

As both the searching and timestamps are small tickets, this seems much simpler.