fix(frontend): Make the fire forecast image size correctly
I've added code to request the WMS image as the right size, namely
setting the bbox to cover Europe, not just the AOI the user has set.
This still isn't great, as we only have the Europe bbox, the raster
image should be in EPSG:3587 (WGS84 Web Mercator), rather than
EPSG:4326 (WGS84), these are subtly different SRS's, we are fixing
in code, but should probably be a fix in GeoServer. Another part
of this fix is to set the layer to use LNG/LAT, not LAT/LNG, which
is the default for mapbox maps, that would also be fixed if the data
were in EPSG:3587.