fix(frontend): Default frequency and resolution to null
Changed the payload being dispatched to default both frequency and resolution to null instead
of empty strings.

The reason for not defaulting inside the form itself is because even if they
are initialized to null at first, typing then deleting will turn the value into an empty string
anyway, so catching it in the final payload is a way to force a null value every time.