fix(frontend): Fix display of event in popup
The name of the field in the event had changed from `damage`, to
`estimated_damage`, so changed that to display the value in the
tooltip popup.

The tooltip was never displaying the event data due to the shape
of the data having changed, to use GeoJSON. This was a good thing,
but had unintended consequences. Simply getting the ID from the
properties object instead of directly from the picked info.object
fixed this problem.

IssueID SAFB-220