fix(frontend): Allow multi-feature popup for in-situ cameras
I have added code to show the popup as normal for in-situ cameras,
but extended it with the capability of displaying multiple features
that happen to exist at the same location. This solution is limited
to just the in-situ cameras popup, to do the same for Fire Alerts
etc, we would need to do something similar.

IssueID #