fix(frontend): Update functionality, styles
Switched to using a useEffect instead of multiple function calls to close the legend, as
all of the dependencies are at the top level, so this makes sense.

Also updated styles to contain the legend within the bounds of the map, pinned to the side and
scrollbars for large legends.