fix(frontend): Fix missing data in popup
I've fixed the missing data in the popup, for when the user clicks
the list item. I don't like the code in this area, but the alerts
list and the index file both create layers and I they overwrite
each other. I suspect it works, but not well. The problem was that
the data is already a GeoJSON feature, but in the AlertsList, we are
re-shaping it to a Feature, incorrectly, removing that, fixed the
problem.

I also found a quick fix for using the scroll wheel, in popus, so
fixed that at the same time.

IssueID SAFB-259