fix(frontend): Re-factor recursive function to display On Demand Layer list
Having changed the shape of the data returned from the API (done in the
backend, in another PR), I have updated the frontend code to use this new
shape. I have also extended it, so that leaf nodes are correctly displayed
and function (except for deleting of requests). A leaf node now shows the
title of the layer, along with it's `status`, if there is a message associated
with the status, then it is displayed on hover of the status text/

IssueID SAFB-106