fix(frontend): Validation working
Had to reshape the validation functions going into the PolygonMap, and the handlers inside
the forms, but it is now working as it should across all 3 forms.

There is some minor duplication of code, but this cannot be helped, as the validation must
take place in the PolygonLayer, and when WKT is entered into the form field, also in the
form field, as the two are separate and do not affect one another.

When a polygon is drawn on the map,however, this value is used in leiu of the form field's value.