fix(frontend): Allow re-rendering of map request list
Previously we prevented updating of map requests on refetching due
to polling of the server, causing lots of re-renders. This in practice
though doesn't affect usability, so is unnecessary. It also had the
side-effect of never updating the list when new requests were
added/removed. This isn't good UX, so I have allowed it to re-render
whenever necessary, with the added benefit of the UI updating when
a new map request is added/deleted.

I also added success/failure notifications, to let the user know the
success/failure of the API request.

Describe why you did what you did

IssueID SAFB-106