fix(frontend): Fix pin functionality for group icons
Fixed the pin functionality for the group icons, which turned out to be a pretty tricky change.

It involved updating the base-level PinLayer to recognise the difference between a cluster and a group,
then call the passed getPinColor method in that case.

Then the function in mapHelpers that determines pin color had to be updated to also recognise groups,
and work out if the selected item is in that group by analyising the coordinates of each.

Finally, each of the Chatbot screens had to be updated to find the selected item from the list
of all, then pass the id and coordinates to the layer handler in order to make all of the above possible.