fix(frontend): Add group icon fix to in-situ cameras
Added the fix to allow group icons to update to the in-situ cameras panel. There were some
strange bugs here that greatly slowed down developent, one of which was there being a second
GeoJsonPinLayer in the alertList for some reason, that was resetting the icons as soon one
was clicked. And the other was the cameraList being initialised to an array, even though the
returned data is a featureCollection object.