fix(frontend): Revert some changes, extend functionality
Reverted removing the duplicated GeoJsonPinLayer in in-situ, as it doesn't need to convert to
GeoJson like the helper one does. However, I had to extend the dupliated one to 1. include the pinInfo
required to identify the selected icon, and 2. pass this function through to the list so that both the
list and map are sharing the same GeoJsonLayer.