fix(frontend): In-situ cameras clicks hooked up
After some refactoring, the clicks for in-situ cameras (which work a bit differently from the others)
are working as expected. This involved a bit of juggling between when to use alert ids and camera ids,
but the camera ids seem to be the correct choice for this kind of identification.

Also refactored the PinLayer to be able to differentiate between clusters and groups in the getIconColor function,
and the generic mapHelper to be able to tell if incoming data is already a feature collection before reshaping.