feat(frontend): Add Polygon representing map request area
Added code to show a Polygon representing the map data requested. This
draws the geometry of the map request, which could be any shape. It
then pans the map to center it on the map. We should also zoom the map,
so the user can see it, as it isn't always obvious, but that will be
done in a future commit.

IssueID SAFB-307