feat(frontend): Add ability to zoom to an area
Added a context/provider/hook, so we can get a reference to the
map. Using the hook, we get access to the ref in context, so we
can interrogate the map and deckgl. Using this allows us to calculate
the zoom level to be used when zooming to a feature.

IssueID SAFB-307