fix(frontend): Don't jump back to data when node re-clicked
The way it worked before was that if you loaded an on-demand layer,
then panned away and clicked the node again, it panned the map back
to the data. Instead the desired functionality is to toggle the
layer visible/invisible, so I have removed the code to pan back to
the location.

IssueID SAFB-308