chore(frontend): changed spelling of chatbot.communication status options
Value was hard-coded to "TakenInCharge" when it should be "Taken In
Charge".  These values for the chatbot modules should really come from
the backend - and there _are_ endpoints to retrieve those values.

This PR is just a quick fix to stop things not working until I have a
chance to refactor the code the right way.