fix(frontend): Enable GetFeatureInfo for On-Demand layers
I've updated the code so we can get feature information about a
selected on-demand layer. The GetTimeSeries, call is ignored as
there never seems to be timeseries data, however, this is hard-coded
in the frontend. It might be better to check, if a URL were to exist,
then allow the option, otherwise, hide the option.