fix(frontend): Fix loading of timeseries
The code, as it was, was very close to working, I have tweaked it
ever so slightly.

I set every layer loaded to have it's visibility value set to `false`,
this loads the data, but keeps it hidden. I then added a new `useEffect`
whose only purpose is to set the visibility to `true` if it matches
the current `timestamp`, otherwise it's set to `false`.

To fix some errors being reported due to propTypes library, I renamed
the props parameter to prop, so it didn't report it as an error.

To fix the slider button not moving as the timeseries moved, the value
was always `NaN`, due to the urls being an object and not an array, as
was expected. I therefore create an array of layers, instead of directly
assigning them to the `tileLayers` state and use the length of that
array, to set the limit of the range slider.