feat(frontend): Introduce interpolated strings for feature info text
Added the ability to inject data into the interpolated `feature_string`
defined in the backend against each data layer. They now all come with
an optional `feature_string`, if it exists, we extract data from the
feature info data response into a known string, with parameter substitution,
otherwise we do as it currently does an use the hard-coded string and
inject a value and unit, if one exists.

IssueID SAFB-318