fix(frontend): Handle when featureString exists with no replacement keys
It is possible for a `feature_string` to exist, but not want any data
interpolated into the string, so I've added a check to only execute
the code, if there are keys to values to be injected, otherwise, just
use the string as is.

I also renamed the function to better indicate it's purpose.

IssueID SAFB-318