fix(frontend): Correct setting className on NavLinks
While it did work, the code was setting the className to an object,
with a key of `active`. This is wrong, at least in my opinion and
was showing an error in the browser console. It should really be
checking what tab is active and adding the `active` class to the
appropriate NavLink.