fix(frontend): Fix missing keys on mapped DOM elements
In react, when you map over an array to generate DOM nodes, you need
to give each a key, so they are uniquely identifiable. This was missing
from a number of mapped nodes, so I've now added them.

The widgets mapping removes the error but isn't great as I'm having
to use the `index`, which is a terrible thing to do, but I have no
other means to set the uniqueness.