fix(frontend): UPdate whitelisted check
The existing whitelist check was erroring and therefore bypassing
the `InProgress` check in the request interceptor. If there was an
exception, that doesn't mean false, so the intercepter did not complete.

The check in `isWhitelisted` is also more complex and old skool than
it should be. Initially I as just going to check that the `url` existed
in the `whitelistedUrls` array of strings, but then learned that some
urls passed in had extra parts, so a simple includes wouldn't work,
therefore I look for an entry in the whitelistedUrls array where that
whitelistedUrl exists in the url passed in as a parameter e.g.

does /data/layers/metadata exist in the url /api/data/layers/metadata/123/abc