feat(frontend): Add properties to drawin fireBreak features, separate features in form
Updated the relationship between form and map so that only the newly drawn feature is added
to the currently selected fireBreak's form field, instead of all the data on the map.

Also updated the drawn fireBreak objects to add the selected fireBreak type (Canadair, Helicopter, etc),
so that they can be correctly colored in map.

IssueID #SAFB-353