feat(frontend): refactor forms, update mapSelection field input, extract custom validator to utility
Made some changes that were requested after PR, mostly little things.

Main change here is editing the shared `mapInputOnChange` handler that is used by all of the
On-Demand forms. The reason you could no longer copy/paste WKT data into the field and have
it update was that it was not changed to work with GeoJson like the rest of the forms (because
it is passed in as a prop from outside)

Also extracted a custom form validator that checks for valid WKT to the datalayers utility file,
to prevent duplication.

IssueID #SAFB-353