feat(frontend): Display Get Feature Info data
A request ot change how the Get Feature Info data is arranged and
displayed was made, this work provides the solution. It takes the
key names as labels and the value. It converts JSON strings back
to proper JSON objects and can be as deeply nested as we want,
however, the more deeply nested, the uglier it looks.

IssueID SAFB-318