feat(project): Set node version to use
For those that use version manager nodenv, or any that adheres to
reading `.node-version` files, I am setting the version of NodeJS
to use. This is also defined in the package.json file, but setting
it there doesn't tell you why something failed, this way we are told
you don't have version X.X.X installed, so it is clearer.