feat(frontend): Some Title text
Restores the `isParentOfSelected` function to match the one inside the original Treeview,
and moves the state setter for `setSelectedNode` to outside the if/else block,
as the conditional was causing the state to not be updated on every click, which was causing
the original bug.

Also found the state being set inside a useEffect too, which was having unexpected side-effects.

IssueID #SAFB-399