fix(frontend): Fix missing team members in missions form
When showing the missions form, the list of members belonging to a
team was never shown. Primarily the lookup was failing due to comparing
a string to an integer. I've fixed that now, improving the code ever
so slightly in the naming of the variable, to be explicit in what was
selected.

IssueID SAFB-407