fix(frontend): Cannot paste WKT into form fields
We made a code change earlier in the week to fix pasting in WKT into
the `communications` and `missions` forms. We failed to test the other
forms like Fire and Burned Areas, Post Event Monitoring etc and it
turns out, we broke them.

I've not reverted the code change we previously did and fixed the
bug in both `communications` and `missions`.