feat(frontend): Add WMTS support for loading layers
Added code to manage the loading of operational layers as WMTS tiles.
The timeseries playback is not 100%, some tiles are taking too long
to retrieve and I cannot yet say if that is a problem with our code,
the deck.gl library or the GeoServer we are getting the tiles from.

IssueID SAFB-242