fix(frontend): Add MapStyle Switcher to form maps
It appears we have 2 maps, I previously added the MapStyle switcher
to the BaseMap, so it can change the map style, but didn't realize
that when we have the forms, the map component used changes to this
PolygonMap. Ideally, we should harmonize them, but that is a different
task, this just fixes it quickly.