feat(frontend): Update page components to set viewstate
Firstly we harmonize the maps to to re-use the `viewState` and use
the `setViewState` from the map hook. This allows all maps to re-use
the same state. The rest of the changes are to remove the local state
managing such stuff, as well as resetting the map back to the default
AOI.

IssueID SAFB-420