fix(frontend): Alerts map not panning or showing data
I broke this by switching to a different setViewState function, that
is more generic. There was some code to check the center points didn't
match, if true, then it panned the map, otherwise, it set the info to
display. That code was weird, I've removed a bit of it, but while this
area works the same as chatbot modules, it is coded quite differently,
making it poor.