fix(frontend): Ensure axios initialized after redux store
Axios uses the store to dispatch some actions, however, it was trying
to use the store to initialize itself, before the store was setup,
causing errors. I've now wrapped the axios interceptors in a function
that we call in the index.js file, providing it with the initialized
store, so the order is maintained that the store is ready to be used.