fix(frontend): On-Demand filtering tree nodes throws exception
This is caused by us using the same function, but the data for the
on-demand and operational layers having different attributes, operational
has `text` and on-demand has `title`, this really should be harmonized.
But, to get around this in the short-term, instead of using the same
function, passed in as a param, I've just copied the function into
each data layer component and slightly tweaked it. This is ugly I admit,
but the true fix would take a lot of effort on the backend and frontend.
The reason why it didn't affect anything before is that the On-Demand
tree view was amending the data inline, but you cannot do that with
redux-toolkit. That is a good thing, although it has caused this issue.