fix(frontend): Add async await to test, pass mock `t()` prop
The test was throwing an `act()` warning because of the asynchronous
effects inside the component, so this waits for that to complete before
running the assertions.

Also, the `t()` prop was required, as it is a function that returns a
string value to populate the buttons, so a mocked function was required
in order to test the rendering of te button by name.