chore(frontend): Update frontend dependencies
I wasn't sure if we had the latest so I deleted my `node_modules`
and `yarn.lock` file and re-generated them. It didn't fix the problem
I'm seeing with missing imports, that turned out to be a problem with
my code editor in the end, but it was worthwhile using the latest
dependencies.