fix(frontend): Redirect to login page on Unauthorized API error
When the user session times out, then network requests are failing
but the user still sees the protected pages. Instead, the app should
redirect to landing page, where they can login again. Before this
wasn't working as the code wasn't removing data in localstorage or
redirecting the browser to the sign-in page, which I've added.