feat(project): Add ability to run tests in CI
Created a github workflow file to set up and run the tests.

I also installed jest-html-reporters to provide pretty reports.
These reports can be found in the action run.

I also added the test output to the .gitignore so that they are
not accidentally comitted.

Removed the "engines" block from package.json, as it was too specific.