fix(frontend): Remove unnecessary view state interactions
There is a lot of passing viewState and setViewState var/functions
in safers. From what I can see, most of it does nothing, or is a
hold-over from when it was passed to the BaseMap. Now that the
BaseMap uses the MapContext, all this is really unnecessary, so I
have removed it.