fix(frontend): Upgrade PolygonMap
A few things going on here:

1. Changed string values to constants and added a transparent color constant.

2. Removed unneeded local state and useEffect for same local state. This was not
required, as the form's `coordinates` prop can just be passed in/out to the same
effect.

3. Increased the stroke width so that drawn features can be seen more easily.

4. Added the fireBreak selecting functionality. I'm not a huge fan of having
something this specific inside a generic Map component, but given lack of time
to refactor, this is the only option, and is completely ignored if not using
fireBreaks, so does not harm other usages of the map.

5. Added a `clearMap` prop and use it inside the map's internal deleting handler,
previously also called `clearMap` but changed to `handleClearMap`. This does nothing
if no feature is selected, but if one is, it will call the passed `clearMap` prop
with the selected feature's data. This allows eash usage of the map to do whatever
logic they need when clearing the map, instead of endless if/else conditions
inside the Map itself.