fix(frontend): Upgrade wildfire simulation form and related files
This upgrades the wildfire form to delete fireBreaks by passing a custom
`clearMap` handler. It also allows the select dropdowns for the fireBreaks
to be governed from both the form and the map, so that the correct fireBreak
is always being edited.

Some constants have been moves, and dataType 35006 has been removed, as per request.

There is also a test suite, and the submit handler for the Wildfire form has been
mopved up to the index, so that it can be passed as a prop for testing purposes.

IssueID #SAFB-301