fix(frontend): Simplify adding Auth Bearer token to requests
Axios has interceptors to do common things to requests, in this situation
we want every request to have the `Authorization: Bearer <token id>`
set. This was done in each request method function, so done in 5 places
rather than the simpler request handler, like it is now. I was also
then able to remove the function used altogether.