fix(frontend): Upgrade outdated dependencies
I've now upgraded to their latest versions, dependencies that were
easy to upgrade. Axios caused a little trouble in that it was being
complained about during jest testing. Adding the transform ignore
flag to tests fixed that.

I also removed `rsuite` and `jest`, the former is not used anywhere
and `jest` is managed by create-react-app, so we shouldn't override
the version it uses.