fix(frontend): Fix Events search this area map button
When clicking this button, at least on the `Events` page, it caused
an error, because the bounding box was not set correctly. This led
to a couple of other changes to get rid of the currentZoomLevel and
midpoint as these are controlled by the viewState. We need to find
all examples of this and remove them.