fix(frontend): Fix Dashboard reference to chatbot report
I didn't realize that the Main dashboard re-used the Report component
so was throwing errors as the prop names had changed. I've now fixed
this, but it is still not quite right. There is the impression that
you can click the report as that is what we want for the list, but
not for the dashboard.