fix(frontend): Refresh app config
Prior to logging in, the app config is fetched and set in redux
global state. However, it is never updated after, meaning you need
to logout/close browser, to get any updates e.g. polling interval.
I've removed the check to see if config is already exists, so that
if you refresh the page, it makes the call again.