feat(backend): refactor oauth2 signin
The first of several PRs to refactor authentication.  This PR removes
the ability to signin locally and only allows signing in via OAuth2.  It
also changes the shape of the authorize response and immediately makes
an additional request to get the user info upon a successful login.  The
goal is to restrict all user state to the "user" object in redux and
leave the "auth" object just for dealing with tokens, etc.