fix(frontend): Fix form validation
The form is not validating correctly when a user selects `citizen`
as their `role`. In this case, there should be no `organization`
selected as that doesn't make sense, but the form could then not be
submitted. I've now fixed the validation to only show the error if
the role is not `citizen`.