fix(frontend): Desctructure thunk params
The thunk takes a single parameter, but what I should have been doing
is extract the `options` attribute from that parameter, not expect
that parameter to be the options itself.

This allows the content of the `options` object to be stringified and
used in the request, rather than trying to stringify and object containing
an object.