fix(frontend): Remove lodash from profile code
Most of the `lodash` usage is to find objects in lists, something
that ES6 introduced long ago, so this extra library, is often obsolete
in a modern project. There are some features that are very useful
and if we use them, fine, but we should prefer natural language features
over lodash alternatives.