fix(frontend): Remove unnecessary form elements
There were a number of form elements that were just not needed, so
I have removed them e.g. Location, City, Address. And some that were
disabled but not actually showing their value, I changed to disabled
text boxes instead, with the actual value visible. I was able to
remove a decent bit of code, which is always good.