feat(frontend): fix error editing events
When an event is edited in the frontend the backend returns an error b/c
the event.id passed as part of the request is undefined.  This is
because the events on the map are GeoJSON and therefore the id is a
feature property rather than being a top-level property.