fix(frontend): Prevent user clicking login button twice
After deleting an account, only from the dashboard API, not from Fusion
Auth as well, the next time you try to login as using that account,
the user was re-created in the Dashboard API, meaning the response is
a 201 Created response, but the frontend code didn't know about that
and didn't see that as a successful login. I've now updated the code
so that if the response is a 200 or 201 status code, then it should
fetch the user's details and complete login, saving the user clicking
the button twice.