fix(frontend): Send refresh token as part of refresh request
When requesting a new `access_token`, we need to send the `refresh_token`
as part of the request, to the server. I get this from the session
data, add it to the request and then update the session data again
with the data from a successful response. I then update the state
with the new `expires_in`, although I don't really see why this is
needed.