fix(frontend): Prevent the loading message when refreshing token
The way the generic code to display a loading message is written,
whenever we make the refresh token request, then the loading message
appears. The user should never even be aware this is happening so in
this case, we don't want the message. However, the way the code has
been written, the current setup won't work, so I had to add a specific
clause, just for the refresh endpoint.