feat(frontend): ability to de-select layer
Code was crashing b/c once a layer was de-selected there is no longer an
object w/ URLs to reference the data from, but the code was still trying
to access those URLs.  Moved that into a ternary that checks if the URLs
(and the entire layer object) exists first.