feat(frontend): remove distinction between local & remote users
Previous versions of the dashboard allowed users to login as a
FusionAuth user or else as a local Django user.  The current version
only supports FusionAuth.  All of the checks in the frontend had already
been removed; This PR just removes the distinction in some of the
mock/testing code.